fix: Ensure motiv icon is always generalised in resource pool
Fixes issue where just adding an additional icon class was not sufficient to reliably and consistently replace original discipline type icon, dure to the inherent cascade order of the FontAwesome styles, since specifcity is equal whatever icon class is lower in the cascade order takes precedence. This is almost all icons in this case. Initial discipline type icon is now completely removed, and then restored if the card is rejected from the resource pool by a card attribute lookup on the discipline type to ensure the correct type is restored.